pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "amcat"
version = "0.1.0"
description = "ANSI Music cat: bit-perfect terminal passthrough with live ANSI-music playback (FluidSynth) and MIDI export"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "amcat contributors"}]
keywords = ["ansi", "bbs", "music", "fluidsynth", "midi", "retro", "cp437"]
requires-python = ">=3.9"
dependencies = ["pyFluidSynth>=1.3", "mido>=1.3"]

[project.scripts]
amcat = "amcat.cli:main"
